Move detector helper file to common
