+++ qt/scientific_interfaces/ISISReflectometry/Common/CMakeLists.txt
@@ -1,9 +1,12 @@
-set(COMMON_SRC_FILES Parse.cpp GetInstrumentParameter.cpp InstrumentParameters.cpp OptionDefaults.cpp Clipboard.cpp)
+set(COMMON_SRC_FILES Detector.cpp Parse.cpp GetInstrumentParameter.cpp InstrumentParameters.cpp OptionDefaults.cpp
+                     Clipboard.cpp
+)
 
 # Include files aren't required, but this makes them appear in Visual Studio IMPORTANT: Include files are required in
 # the MOC_FILES set. Scroll down to find it.
 set(COMMON_INC_FILES
     Clipboard.h
+    Detector.h
     DllConfig.h
     First.h
     GetInstrumentParameter.h
