+++ qt/scientific_interfaces/ISISReflectometry/Common/Detector.cpp
