+++ qt/scientific_interfaces/ISISReflectometry/Common/Detector.h
