+++ qt/scientific_interfaces/ISISReflectometry/GUI/Preview/PreviewModel.h
@@ -44,6 +44,7 @@ public:
   PreviewRow const &getPreviewRow() const override;
 
   void setLoadedWs(Mantid::API::MatrixWorkspace_sptr workspace);
+  void setSummedWs(Mantid::API::MatrixWorkspace_sptr workspace) override;
   void setTheta(double theta) override;
   void setSelectedBanks(std::vector<Mantid::detid_t> selectedBanks) override;
   void setSelectedRegion(ROIType regionType, Selection const &selection) override;
